# src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "servers": [
    {
      "url": "http://10.114.34.65/"
    },
    {
      "url": "https://10.114.34.65/"
    }
  ],
  "paths": {
    "/users/": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/user"
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/user"
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/user"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/user"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "example": "any"
                  },
                  "name": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  },
                  "role_id": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/user"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/user"
                }
              }
            }
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/user"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/user"
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/user"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/user"
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}/img": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/user"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/user"
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/user"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/user"
                }
              }
            }
          }
        }
      }
    },
    "/users/email/free": {
      "post": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/emailFree"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/emailFree"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/roles/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/role"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/role"
                }
              }
            }
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/role"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/role"
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/role"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/role"
                }
              }
            }
          }
        }
      }
    },
...
    "/tags/{tagId}/user/{userId}": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "tagId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ideaWithUser"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ideaWithUser"
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "tagId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ideaNoUser"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ideaNoUser"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "user": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Victor Mike"
          },
          "profile_img": {
            "type": "string",
            "example": ""
          },
          "email": {
            "type": "string",
            "example": "victor.mike@app.com"
          },
          "id": {
            "type": "number",
            "example": 10
          },
          "created_at": {
            "type": "string",
            "example": "2022-11-23T17:23:24.903Z"
          },
          "role": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "example": "Junior DevOps Engineer"
              },
              "id": {
                "type": "number",
                "example": 1
              }
            }
          },
          "comments": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "content": {
                  "type": "string",
                  "example": "Nice idea"
                },
                "id": {
                  "type": "number",
                  "example": 2
                },
                "updated_at": {
                  "type": "string",
                  "example": "2022-11-24T10:36:10.190Z"
                },
                "idea": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 1
                    }
                  }
                },
                "user": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 10
                    },
                    "name": {
                      "type": "string",
                      "example": "Victor Mike"
                    }
                  }
                }
              }
            }
          },
          "ideas": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "title": {
                  "type": "string",
                  "example": "Add coffee machine"
                },
                "content": {
                  "type": "string",
                  "example": "We really should have access to free coffee."
                },
                "created_at": {
                  "type": "string",
                  "example": "2022-11-23T17:52:40.243Z"
                },
                "user": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 10
                    },
                    "name": {
                      "type": "string",
                      "example": "Victor Mike"
                    }
                  }
                },
                "comments": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "content": {
                        "type": "string",
                        "example": "Nice idea"
                      },
                      "id": {
                        "type": "number",
                        "example": 2
                      },
                      "updated_at": {
                        "type": "string",
                        "example": "2022-11-24T10:36:10.190Z"
                      },
                      "idea": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number",
                            "example": 1
                          }
                        }
                      },
                      "user": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number",
                            "example": 10
                          },
                          "name": {
                            "type": "string",
                            "example": "Victor Mike"
                          }
                        }
                      }
                    }
                  }
                },
                "likes": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "idea_id": {
                        "type": "number",
                        "example": 1
                      }
                    }
                  }
                },
                "tags": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "tag": {
                        "type": "object",
                        "properties": {
                          "name": {
                            "type": "string",
                            "example": "Cafeteria"
                          },
                          "id": {
                            "type": "number",
                            "example": 15
                          },
                          "description": {
                            "type": "string",
                            "example": "Ideas related to food"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "likes": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "idea_id": {
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        },
        "xml": {
          "name": "user"
        }
      },
...
      "ideaWithUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "Add coffee machine"
          },
          "description": {
            "type": "string",
            "example": "Ideas related to snacks"
          },
          "users": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 10
                },
                "name": {
                  "type": "string",
                  "example": "Victor Mike"
                }
              }
            }
          }
        },
        "xml": {
          "name": "ideaWithUser"
        }
      },
...
      "role": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Junior DevOps Engineer"
          },
          "id": {
            "type": "number",
            "example": 1
          }
        },
        "xml": {
          "name": "role"
        }
      },
...
      "emailFree": {
        "type": "object",
        "properties": {
          "free": {
            "type": "function"
          }
        },
        "xml": {
          "name": "emailFree"
        }
      }
    }
  }
}